feat: remove setdurabilitypolicy function
